Chatbot-Meets-Chatbot GUI
I've made a GUI version of Chatbot-Meets-Chatbot! It has all of the features of the command-line version (including text output in real time), but with buttons for:

- Begin Conversation
- End Conversation
- Clear
- Export into a .txt file

And option fields for:

- Starting phrase
- Number of loops (<0 is infinite, as it is in the command-line version)
- Bot type options for bots 1 & 2 (Cleverbot, Jabberwackey, & Pandorabot)

Also, I'd like to note that Jabberwackey and Pandorabot appear to not be working correctly. I think this is just an error carried over form the command-line version, as the code for parsing and outputting the bot text is pretty much copied from there.